v4l2: Configure encoder output dimension with the same values
We had a V4L2 encoder that was triggering errors because
the output resolution was always kept unchanged even after
negotiating a different valid input resolution using the input caps.
The patch in this merge request fixes that issue by reusing the
input caps to reconfigure the output resolution in the encoder.

This patch comes from Enrique Ocaña Gonzále.