v4l2videodec: fixup for the qtmediaframework
I found the QT would like to flush in beginning of the video,
then the task of the srcpad would be paused twice with that.

The task of the srcpad would only start at the STOPPED state,
so I think I have to stop the task at the end of the sink
flushing.

Signed-off-by: ayaka <ayaka@soulik.info>